Enable "Publish Preview" for Markdown
Depending on what the Preview Modal will be showing in the final version, we might need to pass the 'mode' over so that we can hide items like 'Description' when doing Markdown. As I'm not sure if the Description will be part of the modal or not, I skipped this for now.